This is a change from Float to BigDecimal to get more precision in the % of coverage

+++ src/main/java/io/jenkins/plugins/coverage/targets/CoverageTreeElement.java
@@ -56,11 +56,11 @@ public class CoverageTreeElement implements Serializable {
 
     @Exported
     public float getNumerator() {
-        return ratio.numerator;
+        return ratio.numerator.floatValue();
     }
 
     @Exported
     public float getDenominator() {
-        return ratio.denominator;
+        return ratio.denominator.floatValue();
     }
 }
